+++ core/memstore/lifetime.go
@@ -1,10 +1,19 @@
-package sessions
+package memstore
 
 import (
 	"sync"
 	"time"
+)
 
-	"github.com/kataras/iris/v12/context"
+var (
+	// Clock is the default clock to get the current time,
+	// it can be used for testing purposes too.
+	//
+	// Defaults to time.Now.
+	Clock func() time.Time = time.Now
+
+	// ExpireDelete may be set on Cookie.Expire for expiring the given cookie.
+	ExpireDelete = time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)
 )
 
 // LifeTime controls the session expiration datetime.
@@ -17,18 +26,29 @@ type LifeTime struct {
 	time.Time
 	timer *time.Timer
 
+	// StartTime holds the Now of the Begin.
+	Begun time.Time
+
 	mu sync.RWMutex
 }
 
-// Begin will begin the life based on the time.Now().Add(d).
+// NewLifeTime returns a pointer to an empty LifeTime instance.
+func NewLifeTime() *LifeTime {
+	return &LifeTime{}
+}
+
+// Begin will begin the life based on the Clock (time.Now()).Add(d).
 // Use `Continue` to continue from a stored time(database-based session does that).
 func (lt *LifeTime) Begin(d time.Duration, onExpire func()) {
 	if d <= 0 {
 		return
 	}
 
+	now := Clock()
+
 	lt.mu.Lock()
-	lt.Time = time.Now().Add(d)
+	lt.Begun = now
+	lt.Time = now.Add(d)
 	lt.timer = time.AfterFunc(d, onExpire)
 	lt.mu.Unlock()
 }
@@ -36,15 +56,23 @@ func (lt *LifeTime) Begin(d time.Duration, onExpire func()) {
 // Revive will continue the life based on the stored Time.
 // Other words that could be used for this func are: Continue, Restore, Resc.
 func (lt *LifeTime) Revive(onExpire func()) {
-	if lt.Time.IsZero() {
+	lt.mu.RLock()
+	t := lt.Time
+	lt.mu.RUnlock()
+
+	if t.IsZero() {
 		return
 	}
 
-	now := time.Now()
-	if lt.Time.After(now) {
-		d := lt.Time.Sub(now)
+	now := Clock()
+	if t.After(now) {
+		d := t.Sub(now)
 		lt.mu.Lock()
-		lt.timer = time.AfterFunc(d, onExpire)
+		if lt.timer != nil {
+			lt.timer.Stop() // Stop the existing timer, if any.
+		}
+		lt.Begun = now
+		lt.timer = time.AfterFunc(d, onExpire) // and execute on-time the new onExpire function.
 		lt.mu.Unlock()
 	}
 }
@@ -53,7 +81,9 @@ func (lt *LifeTime) Revive(onExpire func()) {
 func (lt *LifeTime) Shift(d time.Duration) {
 	lt.mu.Lock()
 	if d > 0 && lt.timer != nil {
-		lt.Time = time.Now().Add(d)
+		now := Clock()
+		lt.Begun = now
+		lt.Time = now.Add(d)
 		lt.timer.Reset(d)
 	}
 	lt.mu.Unlock()
@@ -62,7 +92,7 @@ func (lt *LifeTime) Shift(d time.Duration) {
 // ExpireNow reduce the lifetime completely.
 func (lt *LifeTime) ExpireNow() {
 	lt.mu.Lock()
-	lt.Time = context.CookieExpireDelete
+	lt.Time = ExpireDelete
 	if lt.timer != nil {
 		lt.timer.Stop()
 	}
@@ -71,15 +101,23 @@ func (lt *LifeTime) ExpireNow() {
 
 // HasExpired reports whether "lt" represents is expired.
 func (lt *LifeTime) HasExpired() bool {
-	if lt.IsZero() {
+	lt.mu.RLock()
+	t := lt.Time
+	lt.mu.RUnlock()
+
+	if t.IsZero() {
 		return false
 	}
 
-	return lt.Time.Before(time.Now())
+	return t.Before(Clock())
 }
 
 // DurationUntilExpiration returns the duration until expires, it can return negative number if expired,
 // a call to `HasExpired` may be useful before calling this `Dur` function.
 func (lt *LifeTime) DurationUntilExpiration() time.Duration {
-	return time.Until(lt.Time)
+	lt.mu.RLock()
+	t := lt.Time
+	lt.mu.RUnlock()
+
+	return t.Sub(Clock())
 }